+++ packages/cli/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 0.1.6
 
 - Resolver improvements
   - Make `extensions` configurable